pyproject.toml
[project]
name = "prk-mcp-clinisight-ai-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "lxml>=6.0.2",
    "mcp[cli]>=1.25.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
